meditriage_backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.12
psycopg2-binary>=2.9
requests>=2.31

# Face Recognition
deepface>=0.0.79
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
